Split APIs and types between S3-shares and shares-base
