+++ backend/dataall/modules/s3_datasets_shares/api/__init__.py
@@ -1,9 +1,8 @@
 from dataall.modules.s3_datasets_shares.api import (
-    input_types,
     mutations,
     queries,
     resolvers,
     types,
 )
 
-__all__ = ['resolvers', 'types', 'input_types', 'queries', 'mutations']
+__all__ = ['resolvers', 'queries', 'mutations', 'types']
