+++ backend/dataall/modules/shares_base/api/__init__.py
@@ -1,0 +1,9 @@
+from dataall.modules.shares_base.api import (
+    input_types,
+    mutations,
+    queries,
+    resolvers,
+    types,
+)
+
+__all__ = ['resolvers', 'types', 'input_types', 'queries', 'mutations']
