+++ frontend/src/modules/Environments/services/index.js
@@ -20,6 +20,7 @@ export * from './listEnvironmentNetworks';
 export * from './listEnvironmentPermissions';
 export * from './removeConsumptionRole';
 export * from './removeGroup';
+export * from './searchEnvironmentDataItems';
 export * from './updateEnvironment';
 export * from './updateGroupEnvironmentPermissions';
 export * from './updateConsumptionRole';
