+++ frontend/src/modules/Shares/services/getShareObject.js
@@ -16,11 +16,6 @@ export const getShareObject = ({ shareUri, filter }) => ({
         rejectPurpose
         userRoleForShareObject
         canViewLogs
-        consumptionData {
-          s3AccessPointName
-          sharedGlueDatabase
-          s3bucketName
-        }
         principal {
           principalId
           principalType
