+++ frontend/src/modules/Worksheets/services/index.js
@@ -2,6 +2,7 @@ export * from './createWorksheet';
 export * from './deleteWorksheet';
 export * from './getWorksheet';
 export * from './listS3DatasetsOwnedByEnvGroup';
+export * from './listS3DatasetsSharedWithEnvGroup';
 export * from './listWorksheets';
 export * from './runAthenaSqlQuery';
 export * from './updateWorksheet';
