+++ frontend/src/modules/Worksheets/views/WorksheetView.js
@@ -33,13 +33,13 @@ import {
   getSharedDatasetTables,
   listDatasetTableColumns,
   listValidEnvironments,
-  searchEnvironmentDataItems,
   useClient
 } from 'services';
 import {
   deleteWorksheet,
   getWorksheet,
   listS3DatasetsOwnedByEnvGroup,
+  listS3DatasetsSharedWithEnvGroup,
   runAthenaSqlQuery,
   updateWorksheet
 } from '../services';
@@ -140,38 +140,20 @@ const WorksheetView = () => {
         );
       }
       response = await client.query(
-        searchEnvironmentDataItems({
+        listS3DatasetsSharedWithEnvGroup({
           environmentUri: environment.environmentUri,
-          filter: {
-            ...Defaults.selectListFilter,
-            uniqueShares: true,
-            itemTypes: 'DatasetTable'
-          }
+          groupUri: team
         })
       );
       if (response.errors) {
         dispatch({ type: SET_ERROR, error: response.errors[0].message });
       }
-      if (response.data.searchEnvironmentDataItems.nodes) {
+      if (response.data.listS3DatasetsSharedWithEnvGroup) {
         sharedWithDatabases =
-          response.data.searchEnvironmentDataItems.nodes.map((d) => ({
-            datasetUri: d.datasetUri,
+          response.data.listS3DatasetsSharedWithEnvGroup?.map((d) => ({
             value: d.datasetUri,
-            label: d.sharedGlueDatabaseName,
-            GlueDatabaseName: d.sharedGlueDatabaseName,
-            environmentUri: d.environmentUri,
-            principalId: d.principalId
+            label: d.sharedGlueDatabaseName
           }));
-        // Remove duplicates based on GlueDatabaseName
-        sharedWithDatabases = sharedWithDatabases.filter(
-          (database, index, self) =>
-            index ===
-            self.findIndex(
-              (d) =>
-                d.GlueDatabaseName === database.GlueDatabaseName &&
-                d.principalId === team
-            )
-        );
       }
       setDatabaseOptions(ownedDatabases.concat(sharedWithDatabases));
       setLoadingDatabases(false);
