+++ frontend/src/services/graphql/Environment/index.js
@@ -5,5 +5,4 @@ export * from './listEnvironmentConsumptionRoles';
 export * from './listEnvironmentGroups';
 export * from './listEnvironments';
 export * from './listValidEnvironments';
-export * from './searchEnvironmentDataItems';
 export * from './getConsumptionRolePolicy';
